Show loading screen while downloading location

+++ Palace/Reader2/Bookmarks/TPPBookLocation+Locator.swift
@@ -38,7 +38,7 @@ extension TPPBookLocation {
   
   convenience init?(href: String,
                     type: String,
-                    time: Float? = nil,
+                    time: Double? = nil,
                     part: Float? = nil,
                     chapter: String? = nil,
                     chapterProgression: Float? = nil,
